Task 148: Customizable Timers
I have successfully implemented custom timers for locked websites. Now
when a admin goes to set an unlockable site they will be prompted to
enter in a valid number of minutes that it will be unlocked for. Then
once the timer has been set that website will be locked for that
duration of time.